% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode a string using the openstreet maps API}
\usage{
geocode(x, geojson=T, sp=F, verbose=F, ...)
}
\arguments{
\item{x}{str, a location name to geocode in the form of a character vector of
length 1.}

\item{geojson}{bool, should the geoJSON information also be retrieved}

\item{sp}{bool, should the geoJSON data be converted to spatial objects}

\item{verbose}{bool, verbose argument passed to readOGR}

\item{...}{, additional parameters passed to readOGR}
}
\value{
A dataframe with the reults from the OSM geotag call.

A dataframe with the results from the OSM geotag call. This dataframe may
include lists of coordinates and spatial objects depending on the parameters
specified
}
\description{
Geocodes a string using the OSM API. Has the ability to return
additional information such as the geoJSON string and converting the geojson
string to a spatial object.
}
\examples{
# Get results for the
mexicoTags <- geocode("Mexico", geojson=TRUE, sp=TRUE)

# plot the first result that seems like its a polygon
firstPoly <- grep("Polygon", mexicoTags$geoJSONtype)[1]
sp::plot(mexicoTags$shape[[firstPoly]])

}
