% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson2SPDF.R
\name{geojson2SPDF}
\alias{geojson2SPDF}
\title{Converts A geoJSON string to a Spatial Object}
\usage{
geojson2SPDF(geoJSONstr, ...)
}
\arguments{
\item{geoJSONstr}{str, a valid geoJSON string in a character vector of
length 1.}

\item{...}{, additional parameters passed to readOGR}
}
\value{
A spatial object with class coresponding to the class of the
equivilent geoJSON object type.

A spatial object with class coresponding to the class of the equivilent
geoJSON object type.
}
\description{
Converts a geoJSON string to a spatial object using the
readOGR function from the rgdal package.
}
\examples{
# Make a simple polygon object
simplePolygon <- '{
"type": "Polygon",
"coordinates": [
    [
        [-64.73, 32.31],
        [-80.19, 25.76],
        [-66.09, 18.43],
        [-64.73, 32.31]
        ]
    ]
}'

# transform it to spatial polygon and plot
sp::plot(geojson2SPDF(simplePolygon, verbose=FALSE))

}
